Commit projeto quase finalizado

+++ lib/components/details_widget.dart
@@ -1,32 +1,71 @@
 import 'dart:io';
 
+import 'package:agencia_viagens/components/auth_trips.dart';
 import 'package:agencia_viagens/models/details_model.dart';
-import 'package:agencia_viagens/models/trips_model.dart';
-import 'package:agencia_viagens/pages/details_trips.dart';
+import 'package:agencia_viagens/pages/alter_details.dart';
 import '../services/database.dart';
 import 'package:flutter/material.dart';
-import 'package:intl/intl.dart';
-import 'package:path/path.dart';
-import 'package:syncfusion_flutter_datepicker/datepicker.dart';
 
 class DetailWidget extends StatelessWidget {
   final Details detail;
   final DatabasesTrips dbTrips = DatabasesTrips();
+  final Authentication authTrips = Authentication();
   DetailWidget({required this.detail});
+
+  Future<void> _editTodo(BuildContext context) async {
+    bool auth = await Authentication.authentication();
+
+    if (auth) {
+      Navigator.push(
+        context,
+        MaterialPageRoute(
+          builder: (context) => AlterDetails(id: detail.id),
+        ),
+      );
+    }
+  }
+
+  Future<void> _deleteTodo() async {
+    bool auth = await Authentication.authentication();
+    if (auth) {
+      await dbTrips.deleteTrips(detail.id);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.only(bottom: 20),
       child: Dismissible(
-        onDismissed: (DismissDirection dismissDirection) async {
-          await dbTrips.deleteTrips(detail.id);
+        onDismissed: (direction) async {
+          if (direction == DismissDirection.endToStart) {
+            await _editTodo(context);
+            // Implemente a lógica para editar o item
+          } else if (direction == DismissDirection.startToEnd) {
+            await _deleteTodo();
+            // Implemente a lógica para excluir o item
+          }
+
+          print('Falha na autenticação. A exclusão não é permitida.');
         },
         background: Container(
           color: Colors.red,
           alignment: Alignment.center,
           padding: EdgeInsets.only(left: 16),
           child: Icon(Icons.delete, color: Colors.white),
+        ),
+        secondaryBackground: Container(
+          color: Colors.orange,
+          child: const Align(
+            alignment: Alignment.centerRight,
+            child: Padding(
+              padding: const EdgeInsets.only(right: 16.0),
+              child: Icon(
+                Icons.edit,
+                color: Colors.white,
+              ),
+            ),
+          ),
         ),
         key: Key(detail.id.toString()),
         child: ListTile(
